+++ CHANGELOG.md
@@ -9,6 +9,11 @@ Note: In this file, do not use the hard wrap in the middle of a sentence for com
 -->
 
 ## [Unreleased]
+## [0.2.5] - 2025-03-08
+
+- the bot image will stop with SIGINT
+- the server image will wait 5 seconds before exit
+
 ## [0.2.4] - 2025-03-07
 
 - kube play support